==> 02_standardize.ipynb <==
# Standardize

Pull together our different data sources and combine them into a single dataset.

Import Python tools

import pandas as pd

## Department of Streets and Sanitation

Read in the raw data

dss_df = pd.read_excel(
    "input/FOIA_Trees_Planted_2011_thru_2021.xls",
    skipfooter=2,  # The last row is junk and needs to be cut
    parse_dates=["Plant Date"],
    dtype={
        "Ward": str
    }
)

Clean up the columns

dss_columns = {
    "WorkOrderId": "id",
    "Plant Date": "date",
    "Address": "address",
    "Description": "description",
    "Ward": "ward",
    "APS:": "trees",
}

trimmed_dss_df = (
    dss_df[dss_columns.keys()]
    .rename(columns=dss_columns)
    .sort_values("date")
)

Tag the source

trimmed_dss_df['source'] = 'dss'

Filter it down to tree plants, excluding other types of actions

filtered_dss_df = (
    trimmed_dss_df[trimmed_dss_df.description == 'SED_Tree Planting']
    .drop('description', axis=1)
)

Remove duplicate trees using the identifier field

deduped_dss_df = filtered_dss_df.drop_duplicates("id", keep="first")

## Department of Transportation

Read in the two raw data files

cdot_df = pd.read_csv(
    "input/parsed-01.01.2008_12.31.2020311CSR.1.csv",
    parse_dates=["created_date", "activity_date"]
)

cdot2_df = pd.read_csv(
    "input/tabula-11.30.2018-12.31.2020311CSR.2.csv",
    dtype={
        "Ward": str
    }
)

Create a date field based on a fallback scheme. If the activity date is there, use that. If not, use the created date.

cdot_df['date'] = cdot_df.activity_date.mask(pd.isnull, cdot_df['created_date'])

Clean up the columns

cdot_columns = {
    "sr_number": "id",
    "date": "date",
    "address": "address",
    "ward": "ward"
}

trimmed_cdot_df = (
    cdot_df[cdot_columns.keys()]
    .rename(columns=cdot_columns)
    .sort_values("date")
)

Clean up the ward field so it only contains a number

trimmed_cdot_df.ward = trimmed_cdot_df.ward.apply(lambda x: x.replace("-Ward", "").strip() if not pd.isnull(x) else pd.NA)

trimmed_cdot_df.loc[
    trimmed_cdot_df["ward"] == "0",
    'ward'
] = pd.NA

Since each row represents a single tree, set the value as 1.

trimmed_cdot_df['trees'] = 1

Tag the source

trimmed_cdot_df['source'] = 'cdot'

Get rid of really old dates

filtered_cdot_df = trimmed_cdot_df[trimmed_cdot_df.date >= "2008-01-01"]

Calculate an address column for CDOT's second sheet.

cdot2_df['address'] = cdot2_df['Work\nOrder:\nStreet\nDirection'] + " " + cdot2_df['Work Order:\nStreet Name'] + " " + cdot2_df['Work\nOrder:\nStreet Type']

Clean up the columns

cdot2_columns = {
    "Work Order Number": "id",
    "address": "address",
    "Ward": "ward"
}

trimmed_cdot2_df = (
    cdot2_df[cdot2_columns.keys()]
    .rename(columns=cdot2_columns)
)

Set the tree number again.

trimmed_cdot2_df['trees'] = 1

Tag the source again

trimmed_cdot2_df['source'] = 'cdot'

## Parks Department

Read in the raw data

parks_df = pd.read_csv(
    "input/tabula-Park District Tree Plantings.csv"
)

Standardize the columns

parks_columns = {
    "Park": "address",
    "Tree Total": "trees",
    "Year": "year"
}

trimmed_parks_df = (
    parks_df[parks_columns.keys()]
    .rename(columns=parks_columns)
)

Tag the source

trimmed_parks_df['source'] = 'parks'

Get rid of junky old dates

filtered_parks_df = trimmed_parks_df[trimmed_parks_df.year >= 2008]

## Concatenate

Combine all the dataframes together

concat_df = pd.concat([deduped_dss_df, filtered_cdot_df, trimmed_cdot2_df, filtered_parks_df])

Add a year column

concat_df['year'] = concat_df['date'].dt.year

## Export

Output the results

concat_df.to_csv("./output/trees.csv", index=False)
